refactor(SAVE_FILE): implemented Constants file and refactored loadGame() logic
Moved SAVE_FILE into utils/Constants. Updated loadGame() method to return a Player object, and refactored the loadGame() method to exclude the print statement. Moved the print statement into Main and introduced logic to check that the loaded player is not null before entering the gameplay loop. Implemented simple safeguarding to ensure the gameplay loop does not enter if the loaded player is null.